# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface InterviewSession {
  id: number;
  jobDescription: string;
  resume: string;
  createdAt: string;
}

export interface QAEntry {
  id: number;
  sessionId: number;
  question: string;
  answer: string;
  aiProvider: string;
  createdAt: string;
}

export interface InterviewSessionWithQA {
  id: number;
  jobDescription: string;
  resume: string;
  createdAt: string;
  qaEntries: QAEntry[];
}

export interface CreateInterviewSessionBody {
  jobDescription: string;
  resume: string;
  customPrompt?: string;
  extraContext?: string;
}

export type GenerateAnswerBodyAiProvider =
  (typeof GenerateAnswerBodyAiProvider)[keyof typeof GenerateAnswerBodyAiProvider];

export const GenerateAnswerBodyAiProvider = {
  openai: "openai",
  gemini: "gemini",
} as const;

export interface GenerateAnswerBody {
  question: string;
  aiProvider: GenerateAnswerBodyAiProvider;
}

export interface AddQAEntryBody {
  question: string;
  answer: string;
  aiProvider: string;
}

export interface InterviewSummaryResponse {
  summary: string;
}

export interface InterviewError {
  error: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export interface GeminiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface GeminiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateGeminiConversationBody {
  title: string;
}

export interface SendGeminiMessageBody {
  content: string;
}

export interface GeminiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: GeminiMessage[];
}

export interface GenerateGeminiImageBody {
  prompt: string;
}

export interface GenerateGeminiImageResponse {
  b64_json: string;
  mimeType: string;
}

export interface GeminiError {
  error: string;
}