# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /interview/sessions:
    get:
      operationId: listInterviewSessions
      tags: [interview]
      summary: List all interview sessions
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InterviewSession"
    post:
      operationId: createInterviewSession
      tags: [interview]
      summary: Create a new interview session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInterviewSessionBody"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewSession"
  /interview/sessions/{id}:
    get:
      operationId: getInterviewSession
      tags: [interview]
      summary: Get session with Q&A log
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session with Q&A log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewSessionWithQA"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewError"
    delete:
      operationId: deleteInterviewSession
      tags: [interview]
      summary: Delete a session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /interview/sessions/{id}/answer:
    post:
      operationId: generateAnswer
      tags: [interview]
      summary: Generate an answer for a question (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateAnswerBody"
      responses:
        "200":
          description: SSE stream of answer chunks
          content:
            text/event-stream: {}
  /interview/sessions/{id}/summary:
    get:
      operationId: getInterviewSummary
      tags: [interview]
      summary: Get full session summary text
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Summary text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewSummaryResponse"
        "404":
          description: Not found
  /interview/sessions/{id}/qa:
    post:
      operationId: addQAEntry
      tags: [interview]
      summary: Save a Q&A entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddQAEntryBody"
      responses:
        "201":
          description: Saved entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QAEntry"
...
components:
  schemas:
...
    InterviewSession:
      type: object
      properties:
        id:
          type: integer
        jobDescription:
          type: string
        resume:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - jobDescription
        - resume
        - createdAt
    InterviewSessionWithQA:
      type: object
      properties:
        id:
          type: integer
        jobDescription:
          type: string
        resume:
          type: string
        createdAt:
          type: string
          format: date-time
        qaEntries:
          type: array
          items:
            $ref: "#/components/schemas/QAEntry"
      required:
        - id
        - jobDescription
        - resume
        - createdAt
        - qaEntries
    QAEntry:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        question:
          type: string
        answer:
          type: string
        aiProvider:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - question
        - answer
        - aiProvider
        - createdAt
    CreateInterviewSessionBody:
      type: object
      properties:
        jobDescription:
          type: string
        resume:
          type: string
        customPrompt:
          type: string
        extraContext:
          type: string
      required:
        - jobDescription
        - resume
    GenerateAnswerBody:
      type: object
      properties:
        question:
          type: string
        aiProvider:
          type: string
          enum: [openai, gemini]
      required:
        - question
        - aiProvider
    AddQAEntryBody:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
        aiProvider:
          type: string
      required:
        - question
        - answer
        - aiProvider
    InterviewSummaryResponse:
      type: object
      properties:
        summary:
          type: string
      required:
        - summary
    InterviewError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...