# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all interview sessions
 */
export const ListInterviewSessionsResponseItem = zod.object({
  id: zod.number(),
  jobDescription: zod.string(),
  resume: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListInterviewSessionsResponse = zod.array(
  ListInterviewSessionsResponseItem,
);

/**
 * @summary Create a new interview session
 */
export const CreateInterviewSessionBody = zod.object({
  jobDescription: zod.string(),
  resume: zod.string(),
  customPrompt: zod.string().optional(),
  extraContext: zod.string().optional(),
});

/**
 * @summary Get session with Q&A log
 */
export const GetInterviewSessionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetInterviewSessionResponse = zod.object({
  id: zod.number(),
  jobDescription: zod.string(),
  resume: zod.string(),
  createdAt: zod.coerce.date(),
  qaEntries: zod.array(
    zod.object({
      id: zod.number(),
      sessionId: zod.number(),
      question: zod.string(),
      answer: zod.string(),
      aiProvider: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Delete a session
 */
export const DeleteInterviewSessionParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Generate an answer for a question (SSE stream)
 */
export const GenerateAnswerParams = zod.object({
  id: zod.coerce.number(),
});

export const GenerateAnswerBody = zod.object({
  question: zod.string(),
  aiProvider: zod.enum(["openai", "gemini"]),
});

/**
 * @summary Get full session summary text
 */
export const GetInterviewSummaryParams = zod.object({
  id: zod.coerce.number(),
});

export const GetInterviewSummaryResponse = zod.object({
  summary: zod.string(),
});

/**
 * @summary Save a Q&A entry
 */
export const AddQAEntryParams = zod.object({
  id: zod.coerce.number(),
});

export const AddQAEntryBody = zod.object({
  question: zod.string(),
  answer: zod.string(),
  aiProvider: zod.string(),
});

/**
 * @summary List all conversations
 */
export const ListOpenaiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenaiConversationsResponse = zod.array(
  ListOpenaiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateOpenaiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOpenaiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListOpenaiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListOpenaiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenaiMessagesResponse = zod.array(
  ListOpenaiMessagesResponseItem,
);

/**
 * @summary Send a text message and receive a streaming text response
 */
export const SendOpenaiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendOpenaiMessageBody = zod.object({
  content: zod.string(),
});

/**
 * @summary List all conversations
 */
export const ListGeminiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListGeminiConversationsResponse = zod.array(
  ListGeminiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateGeminiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetGeminiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetGeminiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteGeminiConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListGeminiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListGeminiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListGeminiMessagesResponse = zod.array(
  ListGeminiMessagesResponseItem,
);

/**
 * @summary Send a message and receive an AI response (SSE stream)
 */
export const SendGeminiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendGeminiMessageBody = zod.object({
  content: zod.string(),
});