# lib/api-zod/src/generated/types/generateAnswerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateAnswerBodyAiProvider } from "./generateAnswerBodyAiProvider";

export interface GenerateAnswerBody {
  question: string;
  aiProvider: GenerateAnswerBodyAiProvider;
}