# lib/api-zod/src/generated/types/interviewSessionWithQA.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QAEntry } from "./qAEntry";

export interface InterviewSessionWithQA {
  id: number;
  jobDescription: string;
  resume: string;
  createdAt: Date;
  qaEntries: QAEntry[];
}