# lib/api-zod/src/generated/types/qAEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QAEntry {
  id: number;
  sessionId: number;
  question: string;
  answer: string;
  aiProvider: string;
  createdAt: Date;
}